refactor : `Validator` 클래스 삭제 후 검증 코드를 사용하는 클래스로 이동
